+++ Zillow/Managers/Sink.cs
@@ -7,14 +7,14 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using MQTTnet;
-using TwoMQTT.Core.Communication;
+using TwoMQTT.Core.Managers;
 using Zillow.Models.Shared;
 
-namespace Zillow
+namespace Zillow.Managers
 {
-    public class SinkManager : MQTTManager<Resource, Command>
+    public class Sink : MQTTManager<Resource, Command>
     {
-        public SinkManager(ILogger<SinkManager> logger, IOptions<Opts> sharedOpts, IOptions<Models.SinkManager.Opts> opts, ChannelReader<Resource> inputChannel, ChannelWriter<Command> outputChannel) :
+        public Sink(ILogger<Sink> logger, IOptions<Opts> sharedOpts, IOptions<Models.SinkManager.Opts> opts, ChannelReader<Resource> inputChannel, ChannelWriter<Command> outputChannel) :
             base(logger, opts, inputChannel, outputChannel)
         {
             this.sharedOpts = sharedOpts.Value;
