+++ Zillow/Managers/Source.cs
@@ -3,22 +3,20 @@ using System.Threading.Channels;
 using System.Threading.Tasks;
 using System.Threading;
 using System.Linq;
-using TwoMQTT.Core.Communication;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
-using System;
 using System.Net.Http;
 using Zillow.Models.Shared;
-using System.Net;
 using System.Xml;
 using Newtonsoft.Json.Linq;
+using TwoMQTT.Core.Managers;
 
-namespace Zillow
+namespace Zillow.Managers
 {
-    public class SourceManager : HTTPManager<SlugMapping, Resource, Command>
+    public class Source : HTTPManager<SlugMapping, Resource, Command>
     {
-        public SourceManager(ILogger<SourceManager> logger, IOptions<Opts> sharedOpts, IOptions<Models.SourceManager.Opts> opts, ChannelWriter<Resource> outgoing, ChannelReader<Command> incoming, IHttpClientFactory httpClientFactory) :
+        public Source(ILogger<Source> logger, IOptions<Opts> sharedOpts, IOptions<Models.SourceManager.Opts> opts, ChannelWriter<Resource> outgoing, ChannelReader<Command> incoming, IHttpClientFactory httpClientFactory) :
             base(logger, outgoing, incoming, httpClientFactory.CreateClient())
         {
             this.opts = opts.Value;
@@ -55,7 +53,7 @@ namespace Zillow
             }
 
             var results = await Task.WhenAll(tasks);
-            foreach (var result in results)
+            foreach (var result in results.Where(x => x.Ok))
             {
                 this.logger.LogInformation($"Found {result}");
                 await this.outgoing.WriteAsync(Resource.From(result), cancellationToken);
@@ -74,16 +72,22 @@ namespace Zillow
             var baseUrl = "https://www.zillow.com/webservice/GetZestimate.htm";
             var query = $"zws-id={this.opts.ApiKey}&zpid={key.ZPID}";
             var resp = await this.client.GetAsync($"{baseUrl}?{query}", cancellationToken);
+            if (!resp.IsSuccessStatusCode)
+            {
+                return new Models.SourceManager.Response();
+            }
+
             var content = await resp.Content.ReadAsStringAsync();
             var doc = new XmlDocument();
             doc.LoadXml(content);
             string jsonText = JsonConvert.SerializeXmlNode(doc);
             var jsonObj = JObject.Parse(jsonText);
             var response = jsonObj.SelectToken("$.Zestimate:zestimate.response") ?? new JObject();
-            return new Models.SourceManager.Response 
+            return new Models.SourceManager.Response
             {
                 ZPID = key.ZPID,
                 Amount = response.SelectToken(".zestimate.amount")?.Value<decimal>("#text") ?? default,
+                Ok = true,
             };
         }
     }
