+++ Zillow/Models/Shared/Resource.cs
