+++ Zillow/Program.cs
@@ -5,9 +5,8 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.DependencyInjection;
 using Zillow.Models.Shared;
-using System.Reflection;
-using System;
-using System.Linq;
+using TwoMQTT.Core;
+using Zillow.Managers;
 
 namespace Zillow
 {
@@ -15,7 +14,7 @@ namespace Zillow
     {
         static async Task Main(string[] args)
         {
-            if (PrintVersion(args))
+            if (AppVersion.PrintVersion<Program>(args))
             {
                 return;
             }
@@ -40,32 +39,15 @@ namespace Zillow
                     services.AddSingleton<ChannelWriter<Resource>>(x => dataComms.Writer);
                     services.AddSingleton<ChannelReader<Command>>(x => commandComms.Reader);
                     services.AddSingleton<ChannelWriter<Command>>(x => commandComms.Writer);
-                    services.AddHttpClient<SourceManager>();
-                    services.AddHostedService<SourceManager>();
-                    services.AddHostedService<SinkManager>();
+                    services.AddHttpClient<Source>();
+                    services.AddHostedService<Source>();
+                    services.AddHostedService<Sink>();
                 })
                 .ConfigureLogging((hostingContext, logging) => {
                     logging.AddConsole();
                 });
 
             await builder.RunConsoleAsync();
-        }
-
-        static bool PrintVersion(string[] args) 
-        {
-            var param = args?.Skip(1)?.FirstOrDefault() ?? string.Empty;
-            if (param != "version")
-            {
-                return false;
-            }
-
-            var version = Assembly.GetAssembly(typeof(Program))
-                ?.GetName()
-                ?.Version
-                ?.ToString() ?? "0.0.0.0";
-
-            Console.WriteLine($"v{version}");
-            return true;
         }
     }
 }
